monorepo_test/beamformer/pyproject.toml
[tool.poetry]
name = "beamformer"
version = "1.2.0"
description = "CHIME SPS Beamformer"
authors = ["CHIME SPS <chime_spawg@physics.mcgill.>"]
license = "MIT"
readme = "README.md"
exclude = [
    "tests"
]

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
beam-model = {git = "ssh://git@github.com/CHIMEFRB/beam-model"}
rfi-mitigation = {path = "../rfi-mitigation"}
spshuff = {path = "../spshuff"}
sps-common = {path = "../sps-common"}
sps-databases = {path = "../sps-databases"}
astropy = "~5.2"
attrs = "^22.2.0"
numpy = "^1.24.3"
prometheus-client = "^0.17.1"
pytz = "^2023.3"
scipy = "^1.10.1"

[tool.poetry.group.dev.dependencies]
commitizen = "^3.4.0"
mypy = "^1.4.0"
pre-commit = "^3.3.3"
pytest = "^7.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
